+++ src/api/viewsets/project/view.py
@@ -4,16 +4,22 @@ from rest_framework.response import Response
 from api.viewsets.project.model import Project
 from api.viewsets.project.serializer import ProjectSerializer
 from api.viewsets.project.service import ProjectService
+from users.viewsets.user.service import RegisterUserService
 
 
 class ProjectViewSet(viewsets.ModelViewSet):
     queryset = Project.objects.all()
     serializer_class = ProjectSerializer
 
-    @staticmethod
-    def list(request, *args, **kwargs):
-        # TODO FUTURE, for now it is not possible get projects
-        return Response(status=status.HTTP_403_FORBIDDEN)
+    def list(self, request, *args, **kwargs):
+        if request.user.role == 'ADMIN' and request.user.company:
+            users = RegisterUserService().get_users_by_company_id(request.user.company_id)
+            projects = self.get_queryset().filter(user__in=users.values('uuid'))
+        else:
+            projects = self.get_queryset().filter(user=request.user)
+
+        json_projects = ProjectSerializer(projects, many=True).data
+        return Response(json_projects, status=status.HTTP_200_OK)
 
     @staticmethod
     def create(request, *args, **kwargs):