+++ src/users/viewsets/user/service.py
@@ -6,6 +6,7 @@ import jwt
 from typing import Dict, Union
 
 from django.contrib.auth import get_user_model
+from django.db.models import QuerySet
 from django.http import QueryDict
 from jwt import DecodeError, ExpiredSignatureError
 
@@ -32,6 +33,10 @@ class RegisterUserService:
         logging.info(f"UserService: The user with this email {user['email']} and belong to this company id "
                      f"{user['company'].uuid} has been created successfully")
         return self.create_validation_jwt(created_user)
+
+    @staticmethod
+    def get_users_by_company_id(company_id: str) -> QuerySet:
+        return User.objects.filter(company=company_id)
 
     @staticmethod
     def validate_user(user: QueryDict):